Move checks for world bounds to new file in src/util
